dfdc files

+++ RCAIDE/Library/Methods/Propulsors/Converters/Ducted_Fan/write_input_deck.py
@@ -80,7 +80,8 @@ shoo
     V_ref             = ducted_fan.cruise.design_reference_velocity  
     settings_command  = base_settings_command.format(alt,V_ref,V_inf,B,n,RPM,T,geometry_filename) 
  
-    purge_files([geometry_filename])         
+    if not dfdc_object.settings.keep_files:
+        purge_files([geometry_filename])         
     return settings_command
 
 def make_case_command(dfdc_object,case):
